Fixed OS system identification
On EMR nodes, the `etc/lsb-release` file exists, which was causing
the OS to be incorrectly identified as Ubuntu. This causes the setup
script to fail as it uses the wrong executables (`apt-get` instead
of `yum` etc)

This PR uses the info inside `etc/os-release` as that is the standard
way of reporting the distro info. This works as expected on Amazon's
EMR node and on the Ubuntu nodes.