Matching paths, not file names
The assumption seems to be that you'd like to match individual file names. This doesn't work for us, as we want to recursively match any fixes that match the a given prefix.

Changed the logic to match path globs, and added a test that replicates EMR's setup and reproduces the error.